Initialized Lab 3 slides

+++ 2026/iut/sp2/lab02/assets/code/fib.c
@@ -1,4 +1,4 @@
-int fib(int n)
+long long fib(int n)
 {
     if(n == 0) return 0;
     if(n == 1) return 1;
